requirements.txt
APScheduler==3.10.4
arabic-reshaper==3.0.0
asn1crypto==1.5.1
azure-functions==1.23.0
blinker==1.9.0
brotli==1.1.0
brotlicffi==1.1.0.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==45.0.4
cssselect2==0.8.0
flask==3.1.1
fonttools==4.58.4
html5lib==1.1
idna==3.10
itsdangerous==2.2.0
jinja2==3.1.6
lxml==6.0.0
markupsafe==3.0.2
oscrypto==1.3.0
pillow==11.2.1
pycparser==2.22
pydyf==0.11.0
pyhanko-certvalidator==0.27.0
pyhanko==0.29.1
pypdf==5.6.1
pyphen==0.17.2
python-bidi==0.6.6
python-dotenv==1.0.1
pyyaml==6.0.2
reportlab==4.4.2
requests==2.32.4
requests-file==2.1.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
six==1.17.0
svglib==1.5.1
tinycss2==1.4.0
tinyhtml5==2.0.0
typing-extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.5.0
weasyprint==65.1
webencodings==0.5.1
werkzeug==3.1.3
xhtml2pdf==0.2.17
zopfli==0.2.3.post1
